code/histopathological_analysis/requirements.txt
aicspylibczi==3.1.2
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
bokeh==3.4.1
bounded-pool-executor==0.0.3
cellpose==3.0.8
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colorcet==3.1.0
colour-science==0.4.4
comet-ml==3.37.0
comm==0.2.1
configobj==5.0.8
contourpy==1.2.0
cycler==0.12.1
dask==2024.5.0
dask-expr==1.1.0
datashader==0.16.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.2
dulwich==0.21.7
elementpath==4.1.5
email-validator==2.1.0.post1
everett==3.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastcluster==1.2.6
fasteners==0.19
fastjsonschema==2.19.1
fastremap==1.14.1
filelock==3.13.1
fonttools==4.47.0
fqdn==1.5.1
fsspec==2023.12.2
geopandas==1.0.1
h5py==3.10.0
holoviews==1.18.3
huggingface-hub==0.22.2
idna==3.6
imagecodecs==2024.1.1
imageio==2.33.1
importlib_metadata==7.1.0
interpolation==2.2.4
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.2
isoduration==20.11.0
itables==1.6.3
jedi==0.19.1
jgo==1.0.5
Jinja2==3.1.2
joblib==1.3.2
JPype1==1.5.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.4
linkify-it-py==2.0.3
llvmlite==0.41.1
locket==1.0.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multipledispatch==1.0.0
natsort==8.4.0
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook_shim==0.2.3
numba==0.58.1
numcodecs==0.12.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
ome-types==0.3.4
opencv-contrib-python==4.10.0.82
opencv-contrib-python-headless==4.10.0.82
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openslide-python==1.3.1
overrides==7.4.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.0
panel==1.4.2
param==2.1.0
parso==0.8.3
partd==1.4.2
pexpect==4.9.0
Pillow==9.5.0
Pint==0.23
platformdirs==4.1.0
pqdm==0.2.0
prismtoolbox==0.1.1
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pycparser==2.21
pyct==0.5.0
pydantic==1.10.13
pydantic-compat==0.1.2
Pygments==2.17.2
pynndescent==0.5.12
pyogrio==0.9.0
pyparsing==3.1.1
pyproj==3.6.1
python-box==6.1.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pyvips==2.2.1
pyviz_comms==3.0.2
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
regex==2024.5.10
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
roifile==2024.3.20
rpds-py==0.16.2
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.4.2
scipy==1.13.0
scyjava==1.9.1
seaborn==0.11.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentry-sdk==1.39.2
shapely==2.0.4
SimpleITK==2.3.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
statannotations==0.6.0
sympy==1.12
Tempita==0.5.2
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2023.12.9
tiffslide==2.4.0
timm==0.9.16
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
transformers==4.40.2
triton==2.2.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
uc-micro-py==1.0.3
umap-learn==0.5.6
uri-template==1.3.0
urllib3==2.1.0
uuid==1.30
valis-wsi==1.0.4
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.3.3
weightedstats==0.4.1
widgetsnbextension==4.0.10
wrapt==1.16.0
wurlitzer==3.0.3
xarray==2024.3.0
xmlschema==2.5.1
xsdata==24.3.1
xyzservices==2024.4.0
zarr==2.18.0
zipp==3.18.1
